standalone mode catalog.json, types and clean

+++ src/api/middleware/validators/utils/index.ts
@@ -1,21 +1,23 @@
 import { findFileInCatalog } from '../../../utils/catalog';
 import { MissingParamsProps, NamespaceProps } from './props';
 import fetch from 'node-fetch';
+import { logger } from '../../../utils/logs/winston';
+import { redisHandler } from '../../../catalog/redis/connection';
 
 export const purgeData = async (data) => {
     if (data === 'catalog') {
-        await fetch(`${ process.env.NGINX_SERVICE }/purge${ process.env.API_PREFIX }/catalog`);
+        await fetch(`${process.env.NGINX_SERVICE}/purge${process.env.API_PREFIX}/catalog`);
     }
     if (data && data.length && typeof data[0] === 'object') {
-        for ( const file of data ) {
-            await fetch(`${ process.env.NGINX_SERVICE }/purge${ process.env.API_PREFIX }/assets/media/original${ file.unique_name }`);
-            await fetch(`${ process.env.NGINX_SERVICE }/purge${ process.env.API_PREFIX }/assets/media/full${ file.unique_name }`);
+        for (const file of data) {
+            await fetch(`${process.env.NGINX_SERVICE}/purge${process.env.API_PREFIX}/assets/media/original${file.unique_name}`);
+            await fetch(`${process.env.NGINX_SERVICE}/purge${process.env.API_PREFIX}/assets/media/full${file.unique_name}`);
         }
     }
 };
 
 export const sendResponse = async ({ res, status, data = null, errors = null, purge = 'false' }) => {
-    if (purge !== 'false') {
+    if (purge !== 'false' && process.env.DELEGATED_STORAGE_METHOD !== 'STANDALONE') {
         await purgeData(purge === 'catalog' ? 'catalog' : data);
     }
     return res.status(status).json({ data, errors }).end();
@@ -30,9 +32,9 @@ export const checkNamespace = ({ namespace }: NamespaceProps): boolean => {
 
 export const checkMissingParam = ({ requiredParams, params }: MissingParamsProps) => {
     const errors = [];
-    for ( const param of requiredParams ) {
+    for (const param of requiredParams) {
         if (!params.hasOwnProperty(param)) {
-            errors.push(`${ param } is required`);
+            errors.push(`${param} is required`);
         }
     }
     return errors;
@@ -41,11 +43,7 @@ export const checkMissingParam = ({ requiredParams, params }: MissingParamsProps
 export const generateUniqueName = (file, body, namespace, toWebp) => {
     return (
         file &&
-        `/${ namespace }/${ body.destination ?
-            `${ body.destination }/` :
-            '' }${ toWebp && [ 'image/jpeg', 'image/png' ].includes(file.mimetype) ?
-            file.filename.split('.')[0] + '.webp' :
-            file.filename }`
+        `/${namespace}/${body.destination ? `${body.destination}/` : ''}${toWebp && ['image/jpeg', 'image/png'].includes(file.mimetype) ? file.filename.split('.')[0] + '.webp' : file.filename}`
     );
 };
 
@@ -53,7 +51,7 @@ export const fileIsTooLarge = async (file, params, method = 'POST') => {
     const { uuid, namespace } = params;
     if (file) {
         if (file.size > 10000000) {
-            const itemFound = method === 'PATCH' && ( await findFileInCatalog(uuid, 'uuid') );
+            const itemFound = method === 'PATCH' && (await findFileInCatalog(uuid, 'uuid'));
             return {
                 filename: file.filename,
                 size: file.size,
